made abstract circulator class, moved all appropriate functions there, made output class output both csv and json
