+++ src/sim/simulation/sim.py
@@ -134,7 +134,7 @@ class GrowingSim(Window):
         self.timestep = timestep
         self.vis = vis
         self.cmap = plt.get_cmap("coolwarm")
-        self.output = Output(self, "20240619_for_geo.csv")
+        self.output = Output(self, "20240619_for_geo.csv", "20240619_for_geo.json")
         self.setup()
 
     def get_root_midpointx(self) -> float:
